requirements.txt
fastapi
openai-whisper
langchain
uvicorn
ffmpeg-binaries
langchain-aws
python-dotenv
langchain-chroma
python-multipart